moved component assigner logic into tc-ui
