+++ demo/tc-ui/lib/assign-component.js
@@ -1,11 +1,9 @@
-const {
-	getEnums,
-	getTypes,
-	getPrimitiveTypes,
-} = require('@financial-times/tc-schema-sdk');
-const Components = require('../templates/components/edit-components');
+const { componentAssigner } = require('@financial-times/tc-ui/server');
+
+const customComponents = {};
 
 const customComponentMap = {
+	Paragraph: 'LargeText',
 	// SystemLifecycle: 'LifecycleStage',
 	// ProductLifecycle: 'LifecycleStage',
 	// ServiceTier,
@@ -14,25 +12,9 @@ const customComponentMap = {
 	// Email,
 };
 
-const assignComponent = itemType => {
-	const primitiveTypes = {
-		...getPrimitiveTypes({ output: 'component' }),
-		...customComponentMap,
-	};
-	const objectTypes = getTypes().map(type => type.name);
-	if (itemType) {
-		if (getEnums()[itemType]) {
-			return Components.Enum;
-		}
-		if (primitiveTypes[itemType]) {
-			return Components[primitiveTypes[itemType]];
-		}
-		if (objectTypes.includes(itemType)) {
-			return Components.Relationship;
-		}
-	}
-
-	return Components.Text;
+module.exports = {
+	assignComponent: componentAssigner({
+		customComponents,
+		customTypeMappings: customComponentMap,
+	}),
 };
-
-module.exports = { assignComponent };
