--- demo/tc-ui/templates/components/edit-components/index.jsx
@@ -1,5 +1,0 @@
-const { primitives } = require('@financial-times/tc-ui/server');
-
-module.exports = {
-	...primitives,
-};
